+++ bsd/sys/sys/ioccom.h
@@ -72,7 +72,7 @@
 #include <sys/cdefs.h>
 
 __BEGIN_DECLS
-int	ioctl(int, unsigned long, ...);
+int	ioctl(int, int, ...);
 __END_DECLS
 
 #endif
