+++ include/osv/socket.hh
@@ -29,7 +29,7 @@ public:
     virtual int read(struct uio *uio, int flags) override;
     virtual int write(struct uio *uio, int flags) override;
     virtual int truncate(off_t len) override;
-    virtual int ioctl(u_long com, void *data) override;
+    virtual int ioctl(int com, void *data) override;
     virtual int poll(int events) override;
     virtual int poll_sync(struct pollfd& pfd, timeout_t timeout) override;
     virtual int stat(struct stat* buf) override;
